extract recipe loading to a gate component
