+++ src/sidebar/activateRecipe/ActivateRecipePanel.module.scss
@@ -37,13 +37,6 @@
   align-items: center;
 }
 
-.recipeName {
-  font-size: 1.5rem;
-  font-weight: 700;
-  color: #5217c6;
-  margin-bottom: 0.25rem;
-}
-
 .footer {
   flex-grow: 0;
   flex-shrink: 0;
